refactor: move fetch_coordinates_from_rsids to independent module
